split up some components
